# internal/handlers/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ShieldNet 360 Access Platform API",
    "description": "HTTP surface for the ZTNA access platform (see docs/architecture.md \u00a72). Covers access requests, reviews, grants, policy drafting / simulation, AI explain / suggest, and inbound SCIM (auto-sync users).\n\nThis spec is regenerated by `scripts/generate-swagger.sh` whenever a handler-level change ships (see CONTRIBUTING.md).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin server. The access API is served by cmd/ztna-api."
    }
  ],
  "paths": {
...
    "/access/requests/{id}": {
      "get": {
        "summary": "Get an access request with its state-history audit trail",
        "tags": [
          "access-requests"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request + history"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
...
}